refactor(config): finish harness posture split
Harvested from PR #3506 by @Hmbown.

Refs #3311.
